Prob3 W1L3 Updated

+++ CS390-2025/src/W1L3_HW_Asgmt/problem3/HotelBooking.java
@@ -1,6 +1,6 @@
 package W1L3_HW_Asgmt.problem3;
 
-public record HotelBooking(String hotelName, int nights, double pricePerNight) {
+public record HotelBooking(String hotelName, int nights, int pricePerNight) {
     public HotelBooking {
         if(hotelName == null || hotelName.length() == 0) {
             throw new IllegalArgumentException("Invalid input");
@@ -13,7 +13,7 @@ public record HotelBooking(String hotelName, int nights, double pricePerNight) {
     }
 
 //    public static void main(String[] args) {
-//        HotelBooking hb1 = new HotelBooking("Marina Bay", 3, 100.00);
-//        System.out.printf("Total cost per night: %.2f%n", hb1.totalCost());
+//        HotelBooking hb1 = new HotelBooking("Holiday Inn", 2, 250);
+//        System.out.printf("Total Hotel Cost: %.2f%n", hb1.totalCost()); // 500.00
 //    }
 }
